+++ src/app/[locale]/settings/_components/role-settings/TargetInformationForm.tsx
@@ -1,7 +1,10 @@
 import { JSX } from 'react';
-import { Box, Button, FormControl, IconButton, MenuItem, Select, Typography } from '@mui/material';
+import { Box, FormControl, MenuItem, Select, Typography } from '@mui/material';
 import { BaSyxRbacRule, rbacRuleTargets } from 'lib/services/rbac-service/RbacRulesService';
 import { useTranslations } from 'next-intl';
+import { WildcardOrStringArrayInput } from 'app/[locale]/settings/_components/role-settings/WildcardOrStringArrayInput';
+import { RoleFormModel } from 'app/[locale]/settings/_components/role-settings/RoleDialog';
+import { Control, Controller, UseFormGetValues } from 'react-hook-form';
 
 type TargetInformationProps = {
     readonly targetInformation: BaSyxRbacRule['targetInformation'];
@@ -66,92 +69,6 @@ export const TargetInformationForm = (props: TargetInformationProps) => {
                 <Typography>{props.targetInformation['@type']}</Typography>
             )}
             {permissions}
-        </Box>
-    );
-};
-
-import { useState } from 'react';
-import { Checkbox, FormControlLabel, TextField } from '@mui/material';
-import ControlPointIcon from '@mui/icons-material/ControlPoint';
-import RemoveCircleOutlineIcon from '@mui/icons-material/RemoveCircleOutline';
-import { Control, Controller, useFieldArray, UseFormGetValues } from 'react-hook-form';
-import { RoleFormModel } from 'app/[locale]/settings/_components/role-settings/RoleDialog';
-
-type WildcardOrStringArrayInputProps = {
-    rule: string;
-    control: Control<RoleFormModel, never>;
-    setValue: (name: string, value: any) => void;
-    initialValue: string | string[];
-};
-export const WildcardOrStringArrayInput = (props: WildcardOrStringArrayInputProps) => {
-    const t = useTranslations('settings');
-    const checkIfWildcard = (value: string | string[]) => {
-        if (value === '*') {
-            return true;
-        }
-        return Array.isArray(value) && value[0] === '*';
-    };
-    const [isWildcard, setIsWildcard] = useState(checkIfWildcard(props.initialValue));
-    // @ts-expect-error we expect an type error here since react-hook-form doesn't support fields for arrays of strings
-    const { fields, append, remove } = useFieldArray({
-        name: `targetInformation.${props.rule}`,
-        control: props.control,
-    });
-    const wildcardValueChanged = (value: boolean) => {
-        setIsWildcard(value);
-        props.setValue(`targetInformation.${props.rule}`, value ? '*' : []);
-    };
-
-    return (
-        <Box mt="1em">
-            <Typography variant="h5">{props.rule}</Typography>
-
-            <FormControlLabel
-                control={<Checkbox checked={isWildcard} onChange={(e) => wildcardValueChanged(e.target.checked)} />}
-                label="Wildcard"
-            />
-            {!isWildcard && (
-                <>
-                    {fields.map((_, idx) => (
-                        <Controller
-                            key={`targetInformation.${props.rule}.${idx}`}
-                            name={`targetInformation.${props.rule}.${idx}`}
-                            control={props.control}
-                            render={({ field }) => (
-                                <Box display="flex" flexDirection="row" mb="1em">
-                                    <TextField
-                                        onChange={field.onChange}
-                                        onBlur={field.onBlur}
-                                        inputRef={field.ref}
-                                        fullWidth
-                                        key={idx}
-                                        variant="outlined"
-                                        placeholder="Enter specific values"
-                                        value={field.value}
-                                    />
-
-                                    <IconButton>
-                                        <RemoveCircleOutlineIcon
-                                            onClick={() => {
-                                                remove(idx);
-                                            }}
-                                        />
-                                    </IconButton>
-                                </Box>
-                            )}
-                        />
-                    ))}
-                    <Button
-                        variant="text"
-                        startIcon={<ControlPointIcon />}
-                        onClick={() => {
-                            append('');
-                        }}
-                    >
-                        {t('roles.addButton')}
-                    </Button>
-                </>
-            )}
         </Box>
     );
 };
